created four new functions, refactored my createBlock function, the aim now is to make this concurrent to speed it up from 5m 11seconds.
I have been working on the "block" part of my blockchain project for a few hours today, I have wrote the following functions:

hashBlock: this function gets all the data about a block and hashes this data.

timestamp: this function tells you the time in Ireland/Dublin.

addBlockToChain: this function adds a block to the previous block that was made.

createGenesisBlock: this function creates the first block for the blockchain.

createBlock: I have refactored my createBlock function to now only take in an amount of blocks to create, I was previously passing in a block but I am not doing that anymore.

and at the end of this I am calling all of my functions within my main method.

below is the time it took to create 100 blocks:
time it took to execute sequentially: 5m11.2780988s. The aim now is to make it concurrent with the use of Go-routines, wait-groups and much more.