fixed hard-coded block.prevHash to now get the last blocks hash also added in some checks around my code
I have fixed the hard-coded block.prevHash within a block. I fixed this by writing a function to get the last blocks hash, this function is called "getLastBlockHash".

I have also added in some checks to check the length of a chain, in some of my functions, where a chain is passed in.

So instead of the block.prevHash being: "00000000000000000000000000000", it is now: block.prevHash = getLastBlockHash(blockData). (I call my function that I just wrote).