wrote some code to make a file and store data in that file, changed the type of a transaction to be of type Transaction no longer an integer array, wrote two new functions and I am using these new functions in my createBlock function.
In my KeyPairs/main.go file I have added in some code to add whatever address is made to a file called "Addresses.txt", whenever this code runs it appends a new address to this file.

In my Blocks/block.go file I have changed the transactions type in the Block struct from an int array to be of type Transaction.

I have wrote two new functions, below is a brief description on what each function does:

createTransaction: this function creates a transaction.

processTransaction: this function processes a transaction, if there are two values in the addresses.txt file it will create a transaction, with the 1st value to be the sender and the 2nd be the receiver and a random value is the amount that is being sent. (I am having some issues with having more than 2 values, THIS IS THE NEXT PART TO WOKR ON)!

I have added to my createBlock function, I am now calling my two new functions that I have created: "createTransaction" and "processTransaction", these are then used to fill in the block.transactions section of a block.