if a file is greater than 2, I change the sender and the receiver and I also changed from \n to 0x once a file is read.
In my Blocks/block.go file I have wrote some code to check is the file is more than two to change the sender and receiver.

In my ReadFile/readfile.go file I have change the split from "\n" to "0x", the \n was actually causing some major problems, the \n was setting the receiver to be empty but the 0x fixes this problem!