I have returned to this project, finally!!!!! It has been 5-6 months since I last worked on this project!
I came back to this project yesterday (13/11/25) and I'm working on it today(14/11/25) and I'll be working on it for the future. I'm super happy that I came back to this project (I'm very very happy!!!).

The most important thing that I did on 13/11/25 was connecting the backend to the frontend.

I am able to make blocks and show the data on the frontend about each block.

So, my block.go file makes the blocks, block.go also runs as a http server (this is used for the frontend), then in the Block.jsx file we can fetch that endpoint and show it on the frontend.