I am commiting the work that I did on friday (14/11/25), on today (16/11/25).
The main part that has changed is that now, I have a div for each block, in the blockchain. You click on a button and it will show the state of the blockchain and when you click it again, it will hide the state of the blockchain.

I did a good bit of refactoring in the block.jsx file (chat-gpt helped me).

I have created a style for how a block looks, this file is called block.css.

in my block.go file I have fixed my code. my code used to only create 5 blocks at a time, this was for the frontend part. but now I removed the constant n = 5 and I just made n be a local variable within the createBlocksForFrontend(...) function.

so, when createBlocksForFrontend function is called you type in how many blocks you want to create for the blockchain, and then those blocks will be created.

what do I do next?

the next part I want to tackle is how my proof of work algorithm works (it works... but not exactly correct). currently each block has a valid proof of work field, but this is simply not true for the blockchain. what I need to do is the following:

1. fix the pow bug.
2. if the proof of work (pow) algorithm returns false, we don't create that block, therefore the frontend will have to reflect this.