Today (16/11/25) I did some more work on my blockchain project. See below for more details:
In my block.jsx file I have added a heading 4 for each block along with the index for each block.

Since I added a hr (a line) in my block component I decided to style it. At the moment it is just the color purple, this might evolve in the future.

In my readfile.go file, I removed the main function because my readfile.go is a ReadFile package, therefore I don't need the main function, I also changed my ReadFile function to now be ReadAddresses (I also changed the code a bit) and I wrote another small but useful function, this function is used for my proof of work algorithm.

In my KeyPairs/main.go file, I removed lots of comments (they were not needed!), I refactored my address(...) function to be now a one liner (it used to be a two lined function). I also removed the main function, this is not needed because this is a KeyPairs package.

In my block.go file, I changed the pow field from false to true (for the genesis block), I called my refactored functions (ReadAddresses & ReadKeywords) my old function was called "ReadFile".

I now only print if the block proof of work field is true and if it's false I don't put those blocks to the frontend.

There were lots of comments at the bottom of my main function, so I decided to remove these.

BUT!
** Because I'm now not showing the blocks with the false proof of work field on the frontend, the transaction sender & receiver is now incorrect! I will work on this tomorrow**