I have fixed the problem that I encountered yesterday (16/11/25). below is a detailed explanation of the problem:
What was the problem / bug?

- The bug was as follows, each time a block is made I call a function called "processTransaction(...)", this function does a few things for processing a transaction. if the file is of length 2 we set the corresponding sender and receiver, otherwise the file is of length n, so I set the corresponding sender and receiver again.

The problem, was, I was creating blocks no matter what the state of the proof of work (pow) field was for each block (and sending them to the frontend). but if the pow field is false, the block should not be made, therefore not shown to the frontend! (this is what I did yesterday (16/11/25)). in the processTransaction(...) function there is a check to if "the file is of length n", this is used for making addresses and add them to the Address.txt file.

BUT, if we are not sending the false blocks to the frontend, why are we writing to the Addresses.txt file? - This was the problem.

How did I fix the problem?

- The fix to the problem is actually very straight forward! Pass in a block to the processTransaction(...) function and check to see if the block pow field is true, if it's true, make addresses and add them to the Addresses.txt file, otherwise the pow field is false, do nothing!

What did I change in block.go?

below are the changes that I did in my block.go file:

1. In some of my functions, I now pass the parameters by pointer not by value. I also changed some of the parameter names. for example (t -> transaction, b -> block).

2. I did some refactoring / tidying up in my createBlocksForFrontend(...) function.

3. In my processTransaction(...) function I have added a new parameter (a block - this is used for checking to see if the block pow field is true, if so create addresses and add them to Addresses.txt).