Removed "App" directory due to incorrect behaviour, fixed the issue by putting the main.go file in the root of the project
In my block.go file I did some changes but they are very minimal.

The main part of this commit is the deletion of the "App" directory. I was trying to separate my code into there own distinct folder / package, this is best practise. But once I did this the behaviour of my code broke. And after doing some analysis I decided to delete the "App" directory to mitigate the issues that I was having.

What were the issues ?

- Each block was having a valid (true) proof of work (pow) field for every block - this is wrong. The reason to why this happened is, my code could not find the "keywords.txt" file, this file is needed for my pow algorithm for each of the blocks.

- Once all of the n blocks were mined, once I looked at the div's on the frontend, the "Sender" field in the transaction was empty. This was because my code could not find the "Addresses.txt" file - this is needed for processing a transaction.

But the fix is easy right ?

- You would like to think if my code cannot find the "keywords.txt" && "Addresses.txt" files just move those files to the "App" directory.

- I tried this and low and behold it did not work!

How did I fix it ?

I thought of the idea of deleting the "App" directory and just having the main.go file in the root of my project and then it will have no problem finding the two .txt files - due to them being in the root of my project.

Summary:

So, now instead of the App (my whole project - so far!) running in the "App" directory, I moved the main.go file to the root of my project, therefore it can find the "keywords.txt" file (this is used for my proof of work algorithm) and the "Addresses.txt" file (this is used for processing transactions for a Sender & Receiver).