Removed 3 functions out of block.go (not needed for now...) and wrote two new unit tests, wrote another helper function and lastly implemented my TestInvalidGenesisBlock unit test
In my block.go file I had 3 functions that were not in use at all. They were used at the very start of the project for the backend only but I don't need them at the moment. I have them saved on my computer locally just in case I need them in the future.

The main part of this commit is I have added 2 new unit tests. The first new test, tests to see if a hashblock is correct - this test is called: "TestValidHashBlock", my other test does the exact opposite, it check to see if a hashblock is not hashed - this test is called: "TestInvalidHashBlock".

I also wrote another helper function, it simply converts a string to a bool.

I also implemented my unit test called: "TestInvalidGenesisBlock". This test was empty yesterday (19/11/25), but I filled out this test also today.