Changed a function from private to public scope, deferring or might not work on POS algo anymore, removed code that is needed for POS, removed duplicate of POW algo
In my block.go file I have changed one of my functions from private scope to now be public scope. This small change has also been updated in my block_test.go file.

The main purpose of this commit is the deferring (or I might not even do it all at anymore!) of my POS algorithm. I have worked on it for 2 days (21/11 && 22/11) and it has drove me insane! So I'm not working on it anymore. Therefore I'm removing these part of my code base for now!

I have removed "MyCurrency/main.go" and "pos/pos.go".

Lastly, I'm removing "proof-of-work (pow)/pow.go". The reason to why I'm removing this is, I already a POW folder for my POW algorithm, this is a duplicate - therefore it's not needed!