Implemented my own block-reward system, brought back my currency (from my pos code...), added to frontend to now show the block-reward for each block that has a valid POW field.
In my block.go file, I have implemented my own block reward system. I have wrote a new function, this function simply computes a block reward. A block reward consists of: Subsidy (an amount of crypto, or a token), TransactionFees (the fee you paid to get your block on the chain) and BlockRewardTotal (the sum of Subsidy and TransactionFees - this is you're reward for constructing that block).

In my hashBlock(...) function, I have added more values to be hashed, because I have a new struct, this struct is called: "BlockReward" in my block struct. The values in the BlockReward struct must be hashed!

I have brought back my currency, this is used for my block rewards.

I have updated my block component (for the frontend), to show the block reward in each block that has a valid POW field.

what is Next ?
My next task is to make my transactions fees fluctuate.

Because right now each block has the same block reward because the transaction fees don't change.