Creating my own wallet system, did some refactoring for: main.go and Blocks/block.go and lastly, I'm going to be making my own Decentralized Autonomous Organization (DAO), this is in the "DAO" folder / package - (this is in it's very very early stages, before I work on my DAO, I need to ensure my wallet is up to scratch and is very good & robust!)
Right this is what has happened in the last 2 days or so:

I tried to implement my own wallet system... the initial idea was to have a .txt file with all of your funds (after many hours of work - I deleted this code!) and I also tried other ways such as appending to the file and then summing all of the values up, but again this didn't work! and lastly I tried to have a single value in the .txt file and then add your new tokens to that value in the file, but again this didn't work at all!

So, I decided to delete all of this work (about 2 days or so of work). But now I am just going to use my own "Wallet" package for my wallet system.

-------------------------
Below are my updates for the updated files:

In my block.go file, I wrote a function called: "setWalletFunds", this sets a wallet to have a value of whatever "tokens" is computed to be. Then I just call this function in my "CreateBlockZero" function, only if the POW field is true!

I have created my DAO package, in this package I have a dao.go file. This file will be used later on to contain my backend logic for my DAO. (The wallet is needed for the DAO, so I need to have my wallet package, very solid before going further into my DAO backend code!).

In my main.go file, I have added my DAO package in, I changed the format of the "questions" slice. When you type 1 it runs my genesis & n other blocks, when you type 2 it runs the DAO (very early stages...).

I have created my very own wallet system, for my blockchain. Currently it does not do much, but at the moment it works... I have also created a test file for my wallet, but I'm yet to write any tests...

What is next ?

- Make my wallet package much more robust, currently its not great...

- Write unit tests, to ensure the wallet works in the correct way. (The wallet is needed for the DAO, so I need to have my wallet package, very solid before going further into my DAO backend code!).