Can now show data in file on backend to the frontend and I'm working on validating user input for their secret seed phrase
I'm now able to send the content in the "file.txt" to the frontend - so the user can see their secret seed phrase. This is done in the "Wallet/words.jsx" file.

I'm now working on, validating to see if the user has typed in the correct seed values, if they have done this correctly they will gain access to their wallet, otherwise they will have to re-enter their secret seed phrase again! This work is done in "Wallet/confirm.jsx" file (it's still in the works, it's not fully complete just yet!)